requirements.txt
streamlit==1.28.1
tensorflow==2.15.0
pillow==10.0.1
numpy==1.24.3
opencv-python-headless==4.8.1.78
